+++ gh-page/test-page.js
@@ -15,6 +15,7 @@
         scrollBeyondLastLine: false,
         minimap: { enabled: false },
         unicodeHighlight: { ambiguousCharacters: false },
+        inlayHints: { enabled: 'offUnlessPressed' },
     });
     Object.assign(globalThis, { editor });
     wikiparse.print = (wikitext, include, stage) => {